initial commit 12

+++ Week6_Gradle/src/test/java/com/example/Selenium/Record/Order.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 public record Order(
         String name,
         int quantity,
-        double price,
+        long price,
         String status,
         LocalDate date_on,
         boolean refunded
